+++ src/dialects/mysql/index.js
@@ -59,9 +59,9 @@ assign(Client_MySQL.prototype, {
 
   // Get a raw connection, called by the `pool` whenever a new
   // connection needs to be added to the pool.
-  acquireRawConnection() {
+  acquireRawConnection(settings) {
     return new Promise((resolver, rejecter) => {
-      const connection = this.driver.createConnection(this.connectionSettings);
+      const connection = this.driver.createConnection(settings);
       connection.on('error', (err) => {
         connection.__knex__disposed = err;
       });