+++ src/dialects/postgres/index.js
@@ -103,10 +103,10 @@ assign(Client_PG.prototype, {
 
   // Get a raw connection, called by the `pool` whenever a new
   // connection needs to be added to the pool.
-  acquireRawConnection() {
+  acquireRawConnection(settings) {
     const client = this;
     return new Promise(function(resolver, rejecter) {
-      const connection = new client.driver.Client(client.connectionSettings);
+      const connection = new client.driver.Client(settings);
       connection.connect(function(err, connection) {
         if (err) {
           return rejecter(err);