+++ src/dialects/sqlite3/index.js
@@ -59,17 +59,14 @@ assign(Client_SQLite3.prototype, {
   },
 
   // Get a raw connection from the database, returning a promise with the connection object.
-  acquireRawConnection() {
+  acquireRawConnection(settings) {
     return new Promise((resolve, reject) => {
-      const db = new this.driver.Database(
-        this.connectionSettings.filename,
-        (err) => {
-          if (err) {
-            return reject(err);
-          }
-          resolve(db);
+      const db = new this.driver.Database(settings.filename, (err) => {
+        if (err) {
+          return reject(err);
         }
-      );
+        resolve(db);
+      });
     });
   },